修复变参函数问题

+++ libc/string.c
@@ -1,5 +1,6 @@
-#include "include/string.h"
+#include <stdarg.h>
 #include "include/stdint.h"
+#include "include/stddef.h"
 
 char *strcpy(char *dst, const char *src) {
   char *p = dst;
@@ -9,15 +10,15 @@ char *strcpy(char *dst, const char *src) {
   return dst;
 }
 
-size_t strlen(const char *str) {
-  size_t size = 0;
+long strlen(const char *str) {
+  long size = 0;
   for (const char *p = str; *p != '\0'; p++) {
     size++;
   }
   return size;
 }
 
-void *memcpy(void *dst, const void *src, size_t size) {
+void *memcpy(void *dst, const void *src, long unsigned size) {
   uint8_t *p = (char *)dst;
   const uint8_t *q = (const uint8_t *)src;
   for (int i = 0; i < size; i++) {
@@ -26,7 +27,7 @@ void *memcpy(void *dst, const void *src, size_t size) {
   return dst;
 }
 
-void *memset(void *dst, int ch, size_t size) {
+void *memset(void *dst, int ch, long unsigned size) {
   uint8_t *p = dst;
   for (long i = 0; i < size; i++) {
     p[i] = (uint8_t)ch;
